[JSC] Extract InlineCacheHandler to its own file and add subclass https://bugs.webkit.org/show_bug.cgi?id=310886 rdar://173499272
Reviewed by Yijia Huang.

Move InlineCacheHandler and CacheType out of InlineCacheCompiler.h/.cpp
into dedicated InlineCacheHandler.h/.cpp files. Replace the
`TrailingArray<..., DataOnlyCallLinkInfo> pattern with an
InlineCacheHandlerWithJSCall subclass that holds a single
DataOnlyCallLinkInfo member, since at most one CallLinkInfo is ever
needed per handler. This eliminates the variable-length allocation
overhead for the common case where no JS call is made, and uses
destroying delete with dynamicDowncast for correct polymorphic
destruction.

SaferCPPExpectations changes are carried over from prior code. Except
the non-virtual destructor but that's safe because we have an overloaded
destroying_delete_t delete operator.

No new tests, no behavior change.

Canonical link: https://commits.webkit.org/310116@main
