+++ Source/JavaScriptCore/bytecode/InlineCacheCompiler.h
@@ -28,6 +28,7 @@
 #if ENABLE(JIT)
 
 #include "AccessCase.h"
+#include "InlineCacheHandler.h"
 #include "JITStubRoutine.h"
 #include "JSFunctionInlines.h"
 #include "MacroAssembler.h"
@@ -45,20 +46,8 @@ class CCallHelpers;
 class CodeBlock;
 class PolymorphicAccess;
 class PropertyInlineCache;
-class InlineCacheHandler;
 
 DECLARE_ALLOCATOR_WITH_HEAP_IDENTIFIER(PolymorphicAccess);
-
-enum class CacheType : int8_t {
-    Unset,
-    GetByIdSelf,
-    GetByIdPrototype,
-    PutByIdReplace,
-    InByIdSelf,
-    Stub,
-    ArrayLength,
-    StringLength,
-};
 
 class AccessGenerationResult {
 public:
@@ -163,124 +152,6 @@ private:
     friend class InlineCacheCompiler;
 
     ListType m_list;
-};
-
-class InlineCacheHandler final : public RefCounted<InlineCacheHandler>, public TrailingArray<InlineCacheHandler, DataOnlyCallLinkInfo> {
-    WTF_MAKE_NONCOPYABLE(InlineCacheHandler);
-    friend class InlineCacheCompiler;
-public:
-    using Base = TrailingArray<InlineCacheHandler, DataOnlyCallLinkInfo>;
-
-    static Ref<InlineCacheHandler> create(Ref<InlineCacheHandler>&&, CodeBlock*, PropertyInlineCache&, Ref<PolymorphicAccessJITStubRoutine>&&, std::unique_ptr<PropertyInlineCacheClearingWatchpoint>&&, unsigned callLinkInfoCount);
-    static Ref<InlineCacheHandler> createPreCompiled(Ref<InlineCacheHandler>&&, CodeBlock*, PropertyInlineCache&, Ref<PolymorphicAccessJITStubRoutine>&&, std::unique_ptr<PropertyInlineCacheClearingWatchpoint>&&, AccessCase&, CacheType);
-
-    CodePtr<JITStubRoutinePtrTag> callTarget() const { return m_callTarget; }
-    CodePtr<JITStubRoutinePtrTag> jumpTarget() const { return m_jumpTarget; }
-
-    void aboutToDie();
-    bool containsPC(void* pc) const
-    {
-        if (!m_stubRoutine)
-            return false;
-
-        uintptr_t pcAsInt = std::bit_cast<uintptr_t>(pc);
-        return m_stubRoutine->startAddress() <= pcAsInt && pcAsInt <= m_stubRoutine->endAddress();
-    }
-
-    CallLinkInfo* callLinkInfoAt(const ConcurrentJSLocker&, unsigned index);
-
-    // If this returns false then we are requesting a reset of the owning PropertyInlineCache.
-    bool visitWeak(VM&);
-
-    void dump(PrintStream&) const;
-
-    static Ref<InlineCacheHandler> createNonHandlerSlowPath(CodePtr<JITStubRoutinePtrTag>);
-
-    void addOwner(CodeBlock*);
-    void removeOwner(CodeBlock*);
-
-    PolymorphicAccessJITStubRoutine* stubRoutine() { return m_stubRoutine.get(); }
-
-    InlineCacheHandler* next() const { return m_next.get(); }
-    void setNext(RefPtr<InlineCacheHandler>&& next)
-    {
-        m_next = WTF::move(next);
-    }
-
-    AccessCase* accessCase() const { return m_accessCase.get(); }
-    void setAccessCase(RefPtr<AccessCase>&& accessCase)
-    {
-        m_accessCase = WTF::move(accessCase);
-    }
-
-    static constexpr ptrdiff_t offsetOfCallTarget() { return OBJECT_OFFSETOF(InlineCacheHandler, m_callTarget); }
-    static constexpr ptrdiff_t offsetOfJumpTarget() { return OBJECT_OFFSETOF(InlineCacheHandler, m_jumpTarget); }
-    static constexpr ptrdiff_t offsetOfNext() { return OBJECT_OFFSETOF(InlineCacheHandler, m_next); }
-    static constexpr ptrdiff_t offsetOfUid() { return OBJECT_OFFSETOF(InlineCacheHandler, m_uid); }
-    static constexpr ptrdiff_t offsetOfStructureID() { return OBJECT_OFFSETOF(InlineCacheHandler, m_structureID); }
-    static constexpr ptrdiff_t offsetOfOffset() { return OBJECT_OFFSETOF(InlineCacheHandler, m_offset); }
-    static constexpr ptrdiff_t offsetOfNewStructureID() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s2.m_newStructureID); }
-    static constexpr ptrdiff_t offsetOfNewSize() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s2.m_newSize); }
-    static constexpr ptrdiff_t offsetOfOldSize() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s2.m_oldSize); }
-    static constexpr ptrdiff_t offsetOfHolder() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s1.m_holder); }
-    static constexpr ptrdiff_t offsetOfGlobalObject() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s1.m_globalObject); }
-    static constexpr ptrdiff_t offsetOfCustomAccessor() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s1.m_customAccessor); }
-    static constexpr ptrdiff_t offsetOfModuleNamespaceObject() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s3.m_moduleNamespaceObject); }
-    static constexpr ptrdiff_t offsetOfModuleVariableSlot() { return OBJECT_OFFSETOF(InlineCacheHandler, u.s3.m_moduleVariableSlot); }
-    static constexpr ptrdiff_t offsetOfCallLinkInfos() { return Base::offsetOfData(); }
-
-    StructureID structureID() const { return m_structureID; }
-    PropertyOffset offset() const { return m_offset; }
-    JSCell* holder() const { return u.s1.m_holder; }
-    size_t newSize() const { return u.s2.m_newSize; }
-    size_t oldSize() const { return u.s2.m_oldSize; }
-    StructureID newStructureID() const { return u.s2.m_newStructureID; }
-
-    CacheType cacheType() const { return m_cacheType; }
-
-    DECLARE_VISIT_AGGREGATE;
-
-    // This returns true if it has marked everything it will ever marked. This can be used as an
-    // optimization to then avoid calling this method again during the fixpoint.
-    template<typename Visitor> void propagateTransitions(Visitor&) const;
-
-private:
-    InlineCacheHandler()
-        : Base(0)
-    {
-        disableThreadingChecks();
-    }
-
-    InlineCacheHandler(Ref<InlineCacheHandler>&&, Ref<PolymorphicAccessJITStubRoutine>&&, std::unique_ptr<PropertyInlineCacheClearingWatchpoint>&&, unsigned callLinkInfoCount, CacheType);
-
-    static Ref<InlineCacheHandler> createSlowPath(VM&, AccessType);
-
-    CodePtr<JITStubRoutinePtrTag> m_callTarget;
-    CodePtr<JITStubRoutinePtrTag> m_jumpTarget;
-    StructureID m_structureID { };
-    PropertyOffset m_offset { invalidOffset };
-    UniquedStringImpl* m_uid { nullptr };
-    union {
-        struct {
-            StructureID m_newStructureID { };
-            unsigned m_newSize { };
-            unsigned m_oldSize { };
-        } s2 { };
-        struct {
-            JSCell* m_holder;
-            JSGlobalObject* m_globalObject;
-            void* m_customAccessor;
-        } s1;
-        struct {
-            JSObject* m_moduleNamespaceObject;
-            WriteBarrierBase<Unknown>* m_moduleVariableSlot;
-        } s3;
-    } u;
-    CacheType m_cacheType { CacheType::Unset };
-    RefPtr<InlineCacheHandler> m_next;
-    RefPtr<PolymorphicAccessJITStubRoutine> m_stubRoutine;
-    RefPtr<AccessCase> m_accessCase;
-    std::unique_ptr<PropertyInlineCacheClearingWatchpoint> m_watchpoint;
 };
 
 ALWAYS_INLINE bool canUseMegamorphicGetByIdExcludingIndex(VM& vm, UniquedStringImpl* uid)
