+++ Source/JavaScriptCore/bytecode/PropertyInlineCache.cpp
@@ -497,23 +497,24 @@ CallLinkInfo* PropertyInlineCache::callLinkInfoAt(const ConcurrentJSLocker& lock
     if (auto* handlerIC = dynamicDowncast<HandlerPropertyInlineCache>(*this)) {
         if (handlerIC->m_inlinedHandler) {
             if (handlerIC->m_inlinedHandler->accessCase() == &accessCase)
-                return handlerIC->m_inlinedHandler->callLinkInfoAt(locker, 0);
+                return downcast<InlineCacheHandlerWithJSCall>(*handlerIC->m_inlinedHandler).callLinkInfo(locker);
         }
 
         if (auto* cursor = m_handler.get()) {
             while (cursor) {
                 if (cursor->accessCase() == &accessCase)
-                    return cursor->callLinkInfoAt(locker, 0);
+                    return downcast<InlineCacheHandlerWithJSCall>(*cursor).callLinkInfo(locker);
                 cursor = cursor->next();
             }
         }
         return nullptr;
     }
 
-    // Repatching IC path
     if (!m_handler)
         return nullptr;
-    return m_handler->callLinkInfoAt(locker, index);
+    if (!m_handler->stubRoutine())
+        return nullptr;
+    return m_handler->stubRoutine()->callLinkInfoAt(locker, index);
 }
 
 PropertyInlineCacheSummary PropertyInlineCache::summary(const ConcurrentJSLocker& locker, VM& vm) const
